Move QualType to its own file, change pointer access API.
